# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StockTrade Pro Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  pending: 'pending',
  suspended: 'suspended',
} as const;

export interface User {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  status: UserStatus;
  balance: number;
  rewardPoints: number;
  referralCount: number;
  memberCode: string;
  mustChangePassword: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UpdateUserRequestRole = typeof UpdateUserRequestRole[keyof typeof UpdateUserRequestRole];


export const UpdateUserRequestRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UpdateUserRequestStatus = typeof UpdateUserRequestStatus[keyof typeof UpdateUserRequestStatus];


export const UpdateUserRequestStatus = {
  active: 'active',
  pending: 'pending',
  suspended: 'suspended',
} as const;

export interface UpdateUserRequest {
  role?: UpdateUserRequestRole;
  status?: UpdateUserRequestStatus;
  balance?: number;
  rewardPoints?: number;
}

export type OrderType = typeof OrderType[keyof typeof OrderType];


export const OrderType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  investment: 'investment',
  purchase: 'purchase',
  copy_trade: 'copy_trade',
  membership_fee: 'membership_fee',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  userId: string;
  userEmail: string;
  userName: string;
  type: OrderType;
  description: string;
  amount: number;
  status: OrderStatus;
  createdAt: string;
}

export type CreateOrderRequestType = typeof CreateOrderRequestType[keyof typeof CreateOrderRequestType];


export const CreateOrderRequestType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  investment: 'investment',
  purchase: 'purchase',
  copy_trade: 'copy_trade',
  membership_fee: 'membership_fee',
} as const;

export interface CreateOrderRequest {
  type: CreateOrderRequestType;
  description: string;
  amount: number;
}

export type UpdateOrderRequestStatus = typeof UpdateOrderRequestStatus[keyof typeof UpdateOrderRequestStatus];


export const UpdateOrderRequestStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface UpdateOrderRequest {
  status?: UpdateOrderRequestStatus;
}

export type CopyTradeRequestSide = typeof CopyTradeRequestSide[keyof typeof CopyTradeRequestSide];


export const CopyTradeRequestSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface CopyTradeRequest {
  /** @minItems 1 */
  userIds: string[];
  symbol: string;
  side: CopyTradeRequestSide;
  quantity: number;
  price: number;
}

export interface CopyTradeResponse {
  ordersCreated: number;
  orders: Order[];
}

export interface StatsOverview {
  totalUsers: number;
  newUsersThisWeek: number;
  totalRevenue: number;
  totalInvested: number;
  pendingTransactions: number;
  activeInvestments: number;
  totalOrders: number;
  pendingOrders: number;
}

export type ListUsersParams = {
search?: string;
status?: string;
role?: string;
};

export type ListOrdersParams = {
search?: string;
status?: string;
userId?: string;
};