# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StockTrade Pro Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestType } from './createOrderRequestType';

export interface CreateOrderRequest {
  type: CreateOrderRequestType;
  description: string;
  amount: number;
}